Merge in KW/kopano-webapp from ~SDANGROSHIYA/kopano-webapp:bugfix/KW-3640-deleting-mail-with-read-only-permission-shows-http-500-error to master

* commit '7a66699a5f85c692cc06caf305857de862e20345':
  Show message box when folder has no delete access.

+++ client/zarafa/common/Actions.js
@@ -717,10 +717,20 @@ Zarafa.common.Actions = {
 			records = [ records ];
 		}
 
+		var parentEntryID = records[0].get('parent_entryid');
+		var folder = container.getHierarchyStore().getFolder(parentEntryID);
+		if (folder && folder.hasDeleteOwnRights() === false) {
+			Ext.MessageBox.show({
+				title : _('Insufficient permissions'),
+				msg : _("You have insufficient privileges to delete items in this folder."),
+				cls: Ext.MessageBox.ERROR_CLS,
+				buttons: Ext.MessageBox.OK
+			});
+			return;
+		}
+
 		// Check if the records are deleted from the todolist
-		var recordsFolderEntryid = records[0].getStore().entryId;
-		var folder = container.getHierarchyStore().getFolder(recordsFolderEntryid);
-		if ( folder && folder.isTodoListFolder() ) {
+		if (folder && folder.isTodoListFolder()) {
 			Zarafa.task.Actions.deleteRecordsFromTodoList(records);
 		} else {
 			this.doDeleteRecords(records, askOcc, softDelete);
